requirements.txt
# Django Core
Django==5.0.6
djangorestframework==3.15.1

# JWT Authentication
djangorestframework-simplejwt==5.3.1

# Database & Environment
dj-database-url==2.1.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1

# Production Web Server
gunicorn==22.0.0
whitenoise==6.6.0

# CORS Middleware
django-cors-headers==4.3.1

# Swagger
setuptools==69.5.1
drf-yasg==1.21.7
